+++ lib/pages/settings/dashboards/dashboard_definition_page.dart
@@ -27,19 +27,22 @@ import 'package:lotti/widgets/journal/entry_tools.dart';
 import 'package:material_design_icons_flutter/material_design_icons_flutter.dart';
 import 'package:multi_select_flutter/multi_select_flutter.dart';
 
-class DashboardDetailPage extends StatefulWidget {
-  const DashboardDetailPage({
+class DashboardDefinitionPage extends StatefulWidget {
+  const DashboardDefinitionPage({
     super.key,
     required this.dashboard,
+    this.formKey,
   });
 
   final DashboardDefinition dashboard;
+  final GlobalKey<FormBuilderState>? formKey;
 
   @override
-  State<DashboardDetailPage> createState() => _DashboardDetailPageState();
+  State<DashboardDefinitionPage> createState() =>
+      _DashboardDefinitionPageState();
 }
 
-class _DashboardDetailPageState extends State<DashboardDetailPage> {
+class _DashboardDefinitionPageState extends State<DashboardDefinitionPage> {
   final TagsService tagsService = getIt<TagsService>();
   final JournalDb _db = getIt<JournalDb>();
   final PersistenceLogic persistenceLogic = getIt<PersistenceLogic>();
@@ -140,6 +143,7 @@ class _DashboardDetailPageState extends State<DashboardDetailPage> {
 
   @override
   Widget build(BuildContext context) {
+    final formKey = widget.formKey ?? _formKey;
     final localizations = AppLocalizations.of(context)!;
     return StreamBuilder<List<MeasurableDataType>>(
       stream: stream,
@@ -194,9 +198,9 @@ class _DashboardDetailPageState extends State<DashboardDetailPage> {
         }).toList();
 
         Future<DashboardDefinition> saveDashboard() async {
-          _formKey.currentState!.save();
-          if (_formKey.currentState!.validate()) {
-            final formData = _formKey.currentState?.value;
+          formKey.currentState!.save();
+          if (formKey.currentState!.validate()) {
+            final formData = formKey.currentState?.value;
 
             final private = formData?['private'] as bool? ?? false;
             final active = formData?['active'] as bool? ?? false;
@@ -281,10 +285,11 @@ class _DashboardDetailPageState extends State<DashboardDetailPage> {
                       child: Column(
                         children: [
                           FormBuilder(
-                            key: _formKey,
+                            key: formKey,
                             autovalidateMode:
                                 AutovalidateMode.onUserInteraction,
                             onChanged: () {
+                              debugPrint('onChaged');
                               setState(() {
                                 dirty = true;
                               });
@@ -543,7 +548,7 @@ class EditDashboardPage extends StatelessWidget {
           return EmptyScaffoldWithTitle(localizations.dashboardNotFound);
         }
 
-        return DashboardDetailPage(
+        return DashboardDefinitionPage(
           dashboard: dashboard,
         );
       },