+++ lib/widgets/app_bar/app_bar_version.dart
@@ -1,8 +1,8 @@
-import 'package:auto_route/auto_route.dart';
 import 'package:flutter/material.dart';
 import 'package:lotti/database/database.dart';
 import 'package:lotti/get_it.dart';
 import 'package:lotti/theme.dart';
+import 'package:lotti/widgets/app_bar/auto_leading_button.dart';
 import 'package:package_info_plus/package_info_plus.dart';
 
 class VersionAppBar extends StatefulWidget with PreferredSizeWidget {
@@ -73,9 +73,7 @@ class _VersionAppBarState extends State<VersionAppBar> {
               ],
             ),
             centerTitle: true,
-            leading: AutoLeadingButton(
-              color: AppColors.entryTextColor,
-            ),
+            leading: const TestDetectingAutoLeadingButton(),
           );
         }
       },