+++ lib/widgets/app_bar/task_app_bar.dart
@@ -1,4 +1,3 @@
-import 'package:auto_route/auto_route.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_fadein/flutter_fadein.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
@@ -7,6 +6,7 @@ import 'package:lotti/database/database.dart';
 import 'package:lotti/get_it.dart';
 import 'package:lotti/theme.dart';
 import 'package:lotti/widgets/app_bar/app_bar_version.dart';
+import 'package:lotti/widgets/app_bar/auto_leading_button.dart';
 import 'package:lotti/widgets/tasks/linked_duration.dart';
 
 class TaskAppBar extends StatelessWidget with PreferredSizeWidget {
@@ -44,9 +44,7 @@ class TaskAppBar extends StatelessWidget with PreferredSizeWidget {
               ),
             ),
             centerTitle: true,
-            leading: AutoLeadingButton(
-              color: AppColors.entryTextColor,
-            ),
+            leading: const TestDetectingAutoLeadingButton(),
           );
         }
 
@@ -77,9 +75,7 @@ class TaskAppBar extends StatelessWidget with PreferredSizeWidget {
               ],
             ),
             centerTitle: true,
-            leading: AutoLeadingButton(
-              color: AppColors.entryTextColor,
-            ),
+            leading: const TestDetectingAutoLeadingButton(),
           );
         }
       },