+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added:
-- Widget tests for dashboard definition (no UI changes)
+- Widget tests for dashboard definition page (no UI changes)
 
 ## [0.8.72] - 2022-06-23
 ### Added: